Rework class RtorrentLowSpaceDriver()
The configuring process (argparsing, configparsing and logging setup) should be left in the constructor.
Run() should be relegated to running the driver logic, once a valid configuration is established;

This is a step in the process of achieving code modularity. config.py should deliver a neat configuration object; driver.py should only deal with torrent rotator logic.