config.Configuration() class return a dictionary object instead of a ConfigParser object.
In fact, this should have been set in PR #6 : Modularize config, but it must be done eventually so here we are.
Since we are only using one section in the configfile ['main'], it makes no sense for config.Configuration() class to return a ConfigParser object. It is much simpler to return a dictionary object.